Error Handling in Admin/Toys
The most significant changes in the code are related to error handling, data population, and user interface enhancements.

In `AdminController.cs`, error handling has been added for database operations. This is done using `try-catch` blocks to catch `DbEntityValidationException` exceptions that may occur when adding, editing, or deleting toys or categories. If an exception is caught, a JSON response is returned with a `success` property set to `false` and a `message` property containing the error message.

In `ToyController.cs`, a loop has been added to populate the `AgeGroupList` property of the `ToyViewModel` with the age groups of the toys in the `ToyList` property.

In `AdminToysScript.js`, error handling has been added for the `fetch` calls in the `saveCategory` and `saveToy` functions. If the server returns a response with a `success` property set to `false`, an alert is displayed with the error message.

In `GalleryScripts.js`, functionality for filtering toys by age group has been added. A dropdown for selecting an age group is added to the filter bar, and the `filterProductsByAgeGroup` function is used to filter the toys based on the selected age group.

In `adminToys.css` and `gallery.css`, styles have been added for the new elements and existing styles have been modified.

In `_Layout.cshtml`, the URLs of the category links in the navigation menu have been changed to point to the `Gallery` action instead of the `Search` action.

In `Gallery.cshtml`, a dropdown for selecting an age group has been added to the filter bar, and the `toyData`, `categoryData`, and `ageGroupData` divs have been updated to use the `ToyList`, `CategoryList`, and `AgeGroupList` properties of the `ToyViewModel`, respectively.

In `Search.cshtml`, the heading that displays the search value has been removed, as this is now handled in `Gallery.cshtml`.