Final fixes
The most significant changes include the addition of new action methods in the `AdminController.cs` and `OrderController.cs` files, updates to the `CartScript.js` and `HomeScript.js` files, and the addition of a new view file `TrackOrder.cshtml`.

1. The `AdminController.cs` file now includes two new action methods: `Index()` and `Toys()`. These methods return the views for the admin index page and the toys page respectively, with the latter also returning a toy view model.

2. The `OrderController.cs` file has been updated with a new action method `TrackOrder()`, which returns the view for the track order page along with an order item view model. The `Cart()` method has also been updated to return the view for the cart page with an order view model.

3. The `CartScript.js` file now includes additional validation for the user form and credit card form. The code for generating a random password for the customer has been relocated within the script. An alert has also been added to notify the user if the cart is empty when the checkout button is clicked.

4. The `HomeScript.js` file has been updated with a new function `addToCart()`, which sends a POST request to the server to add a toy to the cart. The `generateProduct()` function has also been updated.

5. A new view file `TrackOrder.cshtml` has been added. This file displays a table of the user's orders if the user is logged in. If the user is not logged in, a message is displayed prompting the user to log in.

Other changes include the removal of an unnecessary line of code in the `GalleryScripts.js` file, the relocation of the script tag for `CartScript.js` in the `Cart.cshtml` file, the addition of a new link in the navigation bar for the track order page in the `_Layout.cshtml` file, and a change in the connection string for the database in the `Web.config` file. The `ToyStore.csproj` file has also been updated to include the new view file `TrackOrder.cshtml`.